requirements.txt
Django==4.2.5
django-filter==23.2
djangorestframework==3.14.0
django-stubs==4.2.5
mypy==1.6.1
isort==5.12.0
black==23.10.0
pylint-django==2.5.4



